4.0.1

+++ solon-ai-core/src/main/java/org/noear/solon/ai/chat/ModelOptionsAmend.java
@@ -94,7 +94,7 @@ public class ModelOptionsAmend<T extends ModelOptionsAmend, X> {
                             if (val.indexOf('.') < 0) {
                                 options.put(entry.getKey(), Integer.parseInt(val));
                             } else {
-                                options.put(entry.getKey(), Float.parseFloat(val));
+                                options.put(entry.getKey(), Double.parseDouble(val));
                             }
                         } else {
                             options.put(entry.getKey(), val);
