+++ solon-ai-core/src/main/java/org/noear/solon/ai/embedding/EmbeddingOptions.java
@@ -48,7 +48,7 @@ public class EmbeddingOptions {
                             if (val.indexOf('.') < 0) {
                                 options.put(entry.getKey(), Integer.parseInt(val));
                             } else {
-                                options.put(entry.getKey(), Float.parseFloat(val));
+                                options.put(entry.getKey(), Double.parseDouble(val));
                             }
                         } else {
                             options.put(entry.getKey(), val);
